+++ src/client.ts
@@ -7,6 +7,7 @@ import { FlagsResource } from './resources/flags.js'
 import { MerchantResource } from './resources/merchant.js'
 import { PayoutsResource } from './resources/payouts.js'
 import { ProgramsResource } from './resources/programs.js'
+import { WebhooksResource } from './resources/webhooks.js'
 
 export class AgentRef {
   readonly programs: ProgramsResource
@@ -16,6 +17,7 @@ export class AgentRef {
   readonly flags: FlagsResource
   readonly billing: BillingResource
   readonly merchant: MerchantResource
+  readonly webhooks: WebhooksResource
 
   constructor(config?: AgentRefConfig) {
     const http = new HttpClient(config)
@@ -27,5 +29,6 @@ export class AgentRef {
     this.flags = new FlagsResource(http)
     this.billing = new BillingResource(http)
     this.merchant = new MerchantResource(http)
+    this.webhooks = new WebhooksResource(http)
   }
 }